feat: add useMovies hooks && useLocalStorage tests
